Move TradingPremisesService to  app/services/
