+++ test/controllers/businessmatching/updateservice/UpdateServiceDateOfChangeControllerSpec.scala
@@ -41,8 +41,8 @@ import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.mvc.{Result, Results}
 import utils.{AuthorisedFixture, DependencyMocks, GenericTestHelper}
 import play.api.test.Helpers._
-import services.StatusService
-import services.businessmatching.{BusinessMatchingService, TradingPremisesService}
+import services.{StatusService, TradingPremisesService}
+import services.businessmatching.BusinessMatchingService
 import uk.gov.hmrc.http.cache.client.CacheMap
 import uk.gov.hmrc.play.frontend.auth.connectors.AuthConnector
